requirements.txt
black==19.10b0
cffi==1.14.0
click==7.1.1
coreschema==0.0.4
cryptography>=3.2
Django<=3.0.5
django-cors-headers==3.2.1
django-rest-framework==0.1.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
djongo==1.3.3
django-seed==0.2.2
djangorestframework-csv
drf-writable-nested==0.6.1
drf-yasg==1.17.1
Faker==4.1.3
future==0.18.2
gunicorn==20.0.4
MarkupSafe==1.1.1
pandas
pycparser==2.18
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
regex==2020.4.4
requests==2.23.0
simplejson==3.17.0
six==1.14.0
sqlparse==0.2.4
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
whitenoise==5.1.0
djangorestframework-simplejwt
